% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_brackets.R
\name{build_round}
\alias{build_round}
\title{Created match rows for a round}
\usage{
build_round(round, entries)
}
\arguments{
\item{round}{Integer indciating the round of the tournament}

\item{entries}{Integer indicating the number of entries remaining in the
tournament}
}
\value{
A dataframe with 4 fieldws
}
\description{
Created match rows for a round
}
\examples{
#none
}
