% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_brackets.R
\name{draw_singles_entries}
\alias{draw_singles_entries}
\title{Draw a random field of singles players}
\usage{
draw_singles_entries(
  n_players = 8,
  return_effect = 0.1,
  std_serve = 0.5,
  std_return = 0.01
)
}
\arguments{
\item{n_players}{Integer, number of players in the field}

\item{return_effect}{Numeric, the log odds advantage for the returner}

\item{std_return}{The standard deviation of player return effects}

\item{std_player}{The standard deviation of player effects}
}
\value{
A tibble with columns \code{seed}, \code{serve}, and \code{return}
}
\description{
This function generates a field of random players for a singles
tournament.  Each player gets a serve rating and a return rating.  We assume
the serve ratings for each player \eqn{j} are IID  with
\eqn{s_j = p_j ~ N(0, \sigma_p^2).}

The return ratings are given by

\eqn{r_j = p_j + \rho + e_j}

where \eqn{\rho} is the fixed effect that represents the advantage for the
returning player and each \eqn{e_j\sim N(0, \sigma_r^2)} are IID with
\eqn{\sigma_r} as the standard deviation of player return effects.  These
ratings are used by \code{play_singles_match()} to simulate any of the tournament
matches.  Players are seeded based on the sum of their serve and return
ratings.
}
\examples{
#NONE
}
