% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{logistic}
\alias{logistic}
\title{A logistic function to get binomial probability team 1 wins}
\usage{
logistic(a1 = 0, a2 = 0)
}
\arguments{
\item{a1}{Linear value of team 1}

\item{a2}{Linear value of team 2}
}
\value{
The probability that team 1 wins according to a logistic model
}
\description{
A logistic function to get binomial probability team 1 wins
}
\examples{
logistic(1, 0)

}
