% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{play_many_singles_matches}
\alias{play_many_singles_matches}
\title{Simulate a collection of singles matches}
\usage{
play_many_singles_matches(matches = sample_matches, extra_cols = NULL)
}
\arguments{
\item{matches}{A dataframe containing fields corresponding to the arguments
of \code{play_singles_match()} to simulate.}

\item{extra_cols}{An optional string vector naming additional columns to
report in the results}
}
\value{
A tibble of match results
}
\description{
Simulate a collection of singles matches
}
\examples{
play_many_singles_matches(sample_matches)
}
