% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{play_rally}
\alias{play_rally}
\title{Simulate a single rally}
\usage{
play_rally(r1 = 0, r2 = 0, n = 1)
}
\arguments{
\item{r1}{Numeric, log odds rating of player1}

\item{r2}{Numeric, log odds rating of player2}

\item{n}{Integer indicating the number of rallies to simulate}
}
\value{
A vector of integers indicating which player won the simulated
rallies
}
\description{
Simulate a single rally
}
\examples{
play_rally(5, 1, 10)
}
