% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{play_singles_game}
\alias{play_singles_game}
\title{Simulate a singles game}
\usage{
play_singles_game(
  a_serve = c(0, 0),
  a_return = c(a_serve[1] + 0.01, a_serve[2] + 0.01),
  f_score = 11,
  server = NA
)
}
\arguments{
\item{a_serve}{Numeric indicating the logodds rating of each player winning
a rally while serving}

\item{a_return}{Numeric indicating the logodds rating of each player winning
a rally while receiving}

\item{f_score}{Integer indicating the number of points required to win the
game}

\item{server}{Integer indicating which player serves first}
}
\value{
Numeric, 2-d indicating each player's score
}
\description{
Simulate a singles game
}
\examples{
#None
}
