% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_matches}
\alias{sample_matches}
\title{Sample Match Data}
\format{
\subsection{\code{sample_matches}}{

A data frame with 10 rows of artifical singles match data
\describe{
\item{g_max}{Integer, maximum number of games}
\item{f_score}{Integer, minimum number of points to win a game}
\item{serve1}{player1's serve rating}
\item{serve2}{player2's serve rating}
\item{return1}{player1's return rating}
\item{return2}{player2's return rating}
}
}
}
\usage{
sample_matches
}
\description{
Sample Match Data
}
\keyword{datasets}
