% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_elimination.R
\name{single_elim_1x}
\alias{single_elim_1x}
\title{Single elimination tournament simulator}
\description{
Single elimination tournament simulator

Single elimination tournament simulator
}
\examples{

## ------------------------------------------------
## Method `single_elim_1x$new`
## ------------------------------------------------

#NONE
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{field}}{The players in the tournament}

\item{\code{match_results}}{Tibble describing the results from all matches played}

\item{\code{standings}}{Tibble describing the current standings}

\item{\code{round}}{The current round of the tournament}

\item{\code{rounds}}{The number of rounds in the tournament}

\item{\code{N_matches_round}}{The number of matches in the current round}

\item{\code{id_cols}}{Column names to join by when setting the next round of matches}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-single_elimination_singles_tournament-new}{\code{single_elim_1x$new()}}
\item \href{#method-single_elimination_singles_tournament-clone}{\code{single_elim_1x$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_elimination_singles_tournament-new"></a>}}
\if{latex}{\out{\hypertarget{method-single_elimination_singles_tournament-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of the single elimination singles R6 class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_elim_1x$new(entries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entries}}{A tibble with the entries in the tournament}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the single elimination singles R6 class
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#NONE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-single_elimination_singles_tournament-clone"></a>}}
\if{latex}{\out{\hypertarget{method-single_elimination_singles_tournament-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{single_elim_1x$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
