% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{play_singles_match}
\alias{play_singles_match}
\title{Simulate a singles match}
\usage{
play_singles_match(
  a_serve1 = 0,
  a_serve2 = 0,
  a_return1 = a_serve1 + 0.01,
  a_return2 = a_serve2 + 0.01,
  g_max = 3,
  f_score = 11
)
}
\arguments{
\item{g_max}{Integer indciating the maximum number of games in the match}

\item{f_score}{Integer indicating the minimum number of points required to
win a game}

\item{a_serve}{Numeric indicating the logodds rating of each player winning
a rally while serving}

\item{a_return}{Numeric indicating the logodds rating of each player winning
a rally while receiving}
}
\value{
A dataframe summarizing the simulated match
}
\description{
Simulate a singles match
}
\examples{
#none
}
